fix code issues
