+++ src/app/my-messages/community-acceptance-message-card/community-acceptance-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
