+++ src/app/my-messages/community-changed-message-card/community-changed-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
