+++ src/app/my-messages/community-deleted-message-card/community-deleted-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
