+++ src/app/my-messages/community-invitation-message-card/community-invitation-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
