+++ src/app/my-messages/community-invitation-message-card/community-invitation-message-card.component.ts
@@ -1,14 +1,14 @@
 import { Component, OnInit } from '@angular/core';
-import { TodoNotificationComponent } from '../todo-notification/todo-notification.component';
+import { TodoNotificationComponentTrait } from '../todo-notification-component-trait';
 import { CommunityRegistrationService } from '../../shared/community-registration.service';
 import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-community-invitation-message-card',
   templateUrl: './community-invitation-message-card.component.html',
-  styleUrls: ['./community-invitation-message-card.component.scss', '../my-messages.component.scss']
+  styleUrls: ['./community-invitation-message-card.component.scss']
 })
-export class CommunityInvitationMessageCardComponent extends TodoNotificationComponent implements OnInit {
+export class CommunityInvitationMessageCardComponent extends TodoNotificationComponentTrait implements OnInit {
 
   constructor(protected communityRegistrationService: CommunityRegistrationService,
               protected dialog: MatDialog) {
