+++ src/app/my-messages/community-join-request-message-card/community-join-request-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
