+++ src/app/my-messages/community-kick-out-message-card/community-kick-out-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
