+++ src/app/my-messages/community-left-message-card/community-left-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
