+++ src/app/my-messages/community-role-changed-message-card/community-role-changed-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
