--- src/app/my-messages/todo-notification/todo-notification.ts
@@ -1,7 +1,0 @@
-import { CommunityUserRegistrationResponse } from '../../shared/community-user-registration-response';
-import { AbstractNotification } from '../abstract-notification';
-
-export interface TodoNotification extends AbstractNotification {
-  registration: CommunityUserRegistrationResponse;
-  communityName: string;
-}
