+++ src/app/my-messages/welcome-message-card/welcome-message-card.component.scss
@@ -1,0 +1,1 @@
+@import "../my-messages.component.scss";
